requirements.txt
fastapi
uvicorn[standard]
langchain
chromadb
sentence-transformers
pypdf
python-multipart
langchain-google-genai
langchain-community
pymupdf
python-dotenv